Remove redundancy of capi pkg
