+++ src/interfaces/capi/main.go
@@ -1,5 +1,3 @@
-// +build !secure
-
 /*******************************************************************************
  * Copyright 2019 Samsung Electronics All Rights Reserved.
  *
@@ -17,20 +15,11 @@
  *
  *******************************************************************************/
 
-/*
- * Edge Orchestration
- *
- * Edge Orchestration support to deliver distributed service process environment.
- *
- * API version: v1-20190307
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
 // Package main provides C interface for orchestration
 package main
 
 ///*******************************************************************************
-// * Copyright 2019 Samsung Electronics All Rights Reserved.
+// * Copyright 2020 Samsung Electronics All Rights Reserved.
 // *
 // * Licensed under the Apache License, Version 2.0 (the "License");
 // * you may not use this file except in compliance with the License.
@@ -45,24 +34,63 @@ package main
 // * limitations under the License.
 // *
 // *******************************************************************************/
-//#define MAX_SVC_INFO_NUM 3
-//typedef struct {
-//	char* ExecutionType;
-//	char* ExeCmd;
-//} RequestServiceInfo;
-//
-//typedef struct {
-//	char* ExecutionType;
-//	char* Target;
-//} TargetInfo;
-//
-//typedef struct {
-//	char*      Message;
-//	char*      ServiceName;
-//	TargetInfo RemoteTargetInfo;
-//} ResponseService;
+/*
+#include <stdlib.h>
+
+#ifndef __ORCHESTRATION_H__
+#define __ORCHESTRATION_H__
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+#define MAX_SVC_INFO_NUM 3
+typedef struct {
+	char* ExecutionType;
+	char* ExeCmd;
+} RequestServiceInfo;
+
+typedef struct {
+	char* ExecutionType;
+	char* Target;
+} TargetInfo;
+
+typedef struct {
+	char*      Message;
+	char*      ServiceName;
+	TargetInfo RemoteTargetInfo;
+} ResponseService;
+
+typedef char* (*identityGetterFunc)();
+typedef char* (*keyGetterFunc)(char* id);
+
+identityGetterFunc iGetter;
+keyGetterFunc kGetter;
+
+static void setPSKHandler(identityGetterFunc ihandle, keyGetterFunc khandle){
+	iGetter = ihandle;
+	kGetter = khandle;
+}
+
+static char* bridge_iGetter(){
+	return iGetter();
+}
+
+static char* bridge_kGetter(char* id){
+	return kGetter(id);
+}
+#ifdef __cplusplus
+}
+
+#endif
+
+#endif // __ORCHESTRATION_H__
+
+*/
 import "C"
 import (
+	"errors"
 	"flag"
 	"log"
 	"math"
@@ -80,10 +108,12 @@ import (
 
 	"orchestrationapi"
 
+	"restinterface/cipher/dummy"
 	"restinterface/cipher/sha256"
 	"restinterface/client/restclient"
 	"restinterface/internalhandler"
 	"restinterface/route"
+	"restinterface/tls"
 
 	"db/bolt/wrapper"
 )
@@ -115,7 +145,7 @@ var (
 )
 
 //export OrchestrationInit
-func OrchestrationInit() (errCode C.int) {
+func OrchestrationInit() C.int {
 	flag.BoolVar(&flagVersion, "v", false, "if true, print version and exit")
 	flag.BoolVar(&flagVersion, "version", false, "if true, print version and exit")
 	flag.Parse()
@@ -128,8 +158,19 @@ func OrchestrationInit() (errCode C.int) {
 	log.Println(">>> buildTags : ", buildTags)
 	wrapper.SetBoltDBPath(dbPath)
 
+	isSecured := false
+	if buildTags == "secure" {
+		log.Println("Orchestration init with secure option")
+		isSecured = true
+	}
+
 	restIns := restclient.GetRestClient()
-	restIns.SetCipher(sha256.GetCipher(cipherKeyFilePath))
+
+	if isSecured {
+		restIns.SetCipher(dummy.GetCipher(cipherKeyFilePath))
+	} else {
+		restIns.SetCipher(sha256.GetCipher(cipherKeyFilePath))
+	}
 
 	servicemgr.GetInstance().SetClient(restIns)
 
@@ -143,12 +184,17 @@ func OrchestrationInit() (errCode C.int) {
 	orcheEngine = builder.Build()
 	if orcheEngine == nil {
 		log.Fatalf("[%s] Orchestaration initalize fail", logPrefix)
-		return
+		return -1
 	}
 
 	orcheEngine.Start(deviceIDFilePath, platform, executionType)
 
-	restEdgeRouter := route.NewRestRouter()
+	var restEdgeRouter *route.RestRouter
+	if isSecured {
+		restEdgeRouter = route.NewRestRouterWithCerti(certificateFilePath)
+	} else {
+		restEdgeRouter = route.NewRestRouter()
+	}
 
 	internalapi, err := orchestrationapi.GetInternalAPI()
 	if err != nil {
@@ -156,14 +202,20 @@ func OrchestrationInit() (errCode C.int) {
 	}
 	ihandle := internalhandler.GetHandler()
 	ihandle.SetOrchestrationAPI(internalapi)
-	ihandle.SetCipher(sha256.GetCipher(cipherKeyFilePath))
+
+	if isSecured {
+		ihandle.SetCipher(dummy.GetCipher(cipherKeyFilePath))
+		ihandle.SetCertificateFilePath(certificateFilePath)
+	} else {
+		ihandle.SetCipher(sha256.GetCipher(cipherKeyFilePath))
+	}
+
 	restEdgeRouter.Add(ihandle)
 	restEdgeRouter.Start()
 
-	errCode = 0
 	log.Println(logPrefix, "orchestration init done")
 
-	return
+	return 0
 }
 
 //export OrchestrationRequestService
@@ -217,6 +269,34 @@ func OrchestrationRequestService(cAppName *C.char, cSelfSelection C.int, cReques
 	ret.RemoteTargetInfo.Target = C.CString(res.RemoteTargetInfo.Target)
 
 	return ret
+}
+
+type customPSKHandler struct{}
+
+func (cHandler customPSKHandler) GetIdentity() string {
+	var cIdentity *C.char
+	cIdentity = C.bridge_iGetter()
+	identity := C.GoString(cIdentity)
+	return identity
+}
+
+func (cHandler customPSKHandler) GetKey(id string) ([]byte, error) {
+	var cKey *C.char
+	cStr := C.CString(id)
+	defer C.free(unsafe.Pointer(cStr))
+
+	cKey = C.bridge_kGetter(cStr)
+	key := C.GoString(cKey)
+	if len(key) == 0 {
+		return nil, errors.New("key is empty")
+	}
+	return []byte(key), nil
+}
+
+//export SetPSKHandler
+func SetPSKHandler(iGetter C.identityGetterFunc, kGetter C.keyGetterFunc) {
+	C.setPSKHandler(iGetter, kGetter)
+	tls.SetPSKHandler(customPSKHandler{})
 }
 
 var count int
